refactor(permanents): Remove slower versions of Rysers and Chin-Huhs permanents. Move _can_calculation_be_performed to the permanents base. Update permanents tests accordingly.
