+++ src/boson_sampling_utilities/permanent_calculators/bs_permanent_calculator_base.py
@@ -1,12 +1,14 @@
 __author__ = "Tomasz Rybotycki"
 
 """
-    This class should be used as a base class for all standard BS permanent calculators. By standard I mean that the
-    matrix and in(out)put states are stored in a variables. It takes care of a lot of boilerplate code.
+    This class should be used as a base class for all standard BS permanent calculators.
+    By standard I mean that the matrix and in(out)put states are stored in a variables.
+    It takes care of a lot of boilerplate code.
 """
 
-from numpy import complex128, ndarray, int64, array, asarray
 from typing import Optional
+
+from numpy import ndarray, int64, array, asarray
 
 from ..permanent_calculators.bs_permanent_calculator_interface import BSPermanentCalculatorInterface
 
@@ -46,4 +48,14 @@ class BSPermanentCalculatorBase(BSPermanentCalculatorInterface):
     @output_state.setter
     def output_state(self, output_state: ndarray) -> None:
         self._output_state = asarray(output_state, dtype=int64)
-        
+
+    def _can_calculation_be_performed(self) -> bool:
+        """
+            Checks if calculation can be performed. For this to happen sizes of given
+            matrix and states have to match.
+
+            :return: Information if the calculation can be performed.
+        """
+        return self._matrix.shape[0] == self._matrix.shape[1] \
+            and len(self._output_state) == len(self._input_state) \
+            and len(self._output_state) == self._matrix.shape[0]
