+++ src/boson_sampling_utilities/permanent_calculators/chin_huh_permanent_calculator.py
@@ -1,21 +1,26 @@
 __author__ = "Tomasz Rybotycki"
 
 """
-    This file holds the implementation of Chin and Huhs method of permanent calculation, as presented in [2].
+    This file holds the implementation of Chin and Huhs method of permanent calculation,
+    as presented in [2].
+    
+    We use 
 """
 
-from typing import List, Optional
+import operator
+from functools import reduce
+from typing import Optional
 
-from numpy import complex128, ndarray
-from scipy.special import binom
+from numpy import complex128, ndarray, nonzero, ones, zeros
 
 from ..permanent_calculators.bs_permanent_calculator_base import BSPermanentCalculatorBase
 
 
 class ChinHuhPermanentCalculator(BSPermanentCalculatorBase):
     """
-        This class is designed to calculate permanent of effective scattering matrix of a boson sampling instance.
-        Note, that it can be used to calculate permanent of given matrix. All that is required that input and output
+        This class is designed to calculate permanent of effective scattering matrix of
+        a boson sampling instance. Note, that it can be used to calculate permanent of
+        given matrix. All that is required that input and output
         states are set to [1, 1, ..., 1] with proper dimensions.
     """
 
@@ -25,63 +30,90 @@ class ChinHuhPermanentCalculator(BSPermanentCalculatorBase):
 
     def compute_permanent(self) -> complex128:
         """
-            This is the main method of the calculator. Assuming that input state, output state and the matrix are
-            defined correctly (that is we've got m x m matrix, and vectors of with length m) this calculates the
-            permanent of an effective scattering matrix related to probability of obtaining output state from given
-            input state.
+            This is the main method of the calculator. Assuming that input state,
+            output state and the matrix are defined correctly (that is we've got m x m
+            matrix, and vectors of with length m) this calculates the permanent of an
+            effective scattering matrix related to probability of obtaining output state
+            from a given input state.
+
             :return: Permanent of effective scattering matrix.
         """
         if not self._can_calculation_be_performed():
             raise AttributeError
 
-        v_vectors = self._calculate_v_vectors()
+        # Required for Guan Code iteration
+        v_vector = zeros(len(self._input_state), dtype=int)  # g
+        code_update_information = ones(len(self._input_state), dtype=int)  # u
+        position_limits = list(self._input_state)  # n
 
         permanent = complex128(0)
-        for v_vector in v_vectors:
-            permanent += self._compute_permanent_addend(v_vector)
-        permanent /= pow(2, sum(self._input_state))
-        return permanent
+        sums = dict()
 
-    def _can_calculation_be_performed(self) -> bool:
-        """
-            Checks if calculation can be performed. For this to happen sizes of given matrix and states have
-            to match.
-            :return: Information if the calculation can be performed.
-        """
-        return self._matrix.shape[0] == self._matrix.shape[1] and len(self._output_state) == len(self._input_state) \
-               and len(self._output_state) == self._matrix.shape[0]
+        binomials_product = 1
+        considered_columns_indices = nonzero(self._output_state)[0]
 
-    def _calculate_v_vectors(self, input_vector: Optional[ndarray] = None) -> List[ndarray]:
-        if input_vector is None:
-            input_vector = []
-        v_vectors = []
-        for i in range(int(self._input_state[len(input_vector)]) + 1):
-            input_state = input_vector.copy()
-            input_state.append(i)
+        multiplier = 1
 
-            if len(input_state) == len(self._input_state):
-                v_vectors.append(input_state)
+        # Initialization (0-th step).
+        for i in considered_columns_indices:
+            sums[i] = 0
+            for j in range(len(self._input_state)):
+                sums[i] += self._input_state[j] * self._matrix[i][j]
+
+        permanent += multiplier * binomials_product * \
+                     reduce(operator.mul, [pow(sums[i], self._output_state[i])
+                                           for i in considered_columns_indices], 1)
+
+        # Rest of the steps.
+        while v_vector[-1] <= position_limits[-1]:
+
+            # UPDATE R VECTOR
+            index_to_update = 0  # i
+            updated_value_at_index = v_vector[0] + code_update_information[0]  # k
+            while updated_value_at_index > position_limits[index_to_update] \
+                    or updated_value_at_index < 0:
+                code_update_information[index_to_update] = -code_update_information[
+                    index_to_update]
+                index_to_update += 1
+
+                if index_to_update == len(v_vector):
+
+                    for _ in range(sum(self._input_state)):
+                        permanent /= 2
+
+                    return permanent
+
+                updated_value_at_index = \
+                    v_vector[index_to_update] + \
+                    code_update_information[index_to_update]
+
+            last_value_at_index = v_vector[index_to_update]
+            v_vector[index_to_update] = updated_value_at_index
+            # END UPDATE
+
+            # START PERMANENT UPDATE
+            multiplier = -multiplier
+
+            # Sums update
+            for i in sums:
+                sums[i] -= 2 * (v_vector[index_to_update] - last_value_at_index) * \
+                           self.matrix[i][index_to_update]
+
+            # Binoms update
+            if v_vector[index_to_update] > last_value_at_index:
+                binomials_product *= (self._input_state[index_to_update] -
+                                      last_value_at_index) / v_vector[index_to_update]
             else:
-                v_vectors.extend(self._calculate_v_vectors(input_state))
+                binomials_product *= last_value_at_index / (
+                        self._input_state[index_to_update] - v_vector[index_to_update])
 
-        return v_vectors
+            addend = multiplier * binomials_product * \
+                     reduce(operator.mul, [pow(sums[j], self._output_state[j])
+                                           for j in considered_columns_indices], 1)
 
-    def _compute_permanent_addend(self, v_vector: ndarray) -> complex128:
-        v_sum = sum(v_vector)
-        addend = pow(-1, v_sum)
-        # Binomials calculation
-        for i in range(len(v_vector)):
-            addend *= binom(self._input_state[i], v_vector[i])
-        # Product calculation
-        product = 1
-        for j in range(len(self._input_state)):
-            if self._output_state[j] == 0:  # There's no reason to calculate the sum if t_j = 0
-                continue
-            # Otherwise we calculate the sum
-            product_part = 0
-            for i in range(len(self._input_state)):
-                product_part += (self._input_state[i] - 2 * v_vector[i]) * self._matrix[j][i]
-            product_part = pow(product_part, self._output_state[j])
-            product *= product_part
-        addend *= product
-        return addend
+            permanent += addend
+
+        for _ in range(sum(self._input_state)):
+            permanent /= 2
+
+        return permanent
