+++ src_tests/__init__.py
@@ -7,8 +7,6 @@ from src.boson_sampling_utilities.permanent_calculators.classic_permanent_calcul
 from src.boson_sampling_utilities.permanent_calculators.bs_permanent_calculator_factory import *
 from src.boson_sampling_utilities.permanent_calculators.ryser_permanent_calculator import *
 from src.boson_sampling_utilities.permanent_calculators.ryser_guan_permanent_calculator import *
-from src.boson_sampling_utilities.permanent_calculators.ryser_guan_permanent_calculator_v2 import *
-from src.boson_sampling_utilities.permanent_calculators.chin_huh_permanent_calculator_v2 import *
 
 from src.distribution_calculators.bs_distribution_calculator_with_fixed_losses import *
 from src.distribution_calculators.bs_exact_distribution_with_uniform_losses import *
